Mistral/requirements.txt
fastapi
uvicorn[standard]
python-multipart
httpx

# Vision
torch
transformers
Pillow

# Mistral API
mistralai
python-dotenv

pydantic
sqlalchemy
alembic

# Frontend
streamlit
requests

# Tests
pytest
pytest-asyncio
ruff
